requirements.txt
pandas
PyMuPDF
configparser
openai
python-docx
